requirements.txt
# Core dependencies
claude-agent-sdk>=0.1.0
anthropic>=0.69.0
python-dotenv>=1.0.0

# Google Sheets integration
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Async and networking
aiohttp>=3.9.0
httpx>=0.25.0

# Data validation and serialization
pydantic>=2.5.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# Development and testing tools (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
flake8>=6.0.0

# Utilities
psutil>=5.9.0
python-dateutil>=2.8.2